accounts-service/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
psycopg2-binary
pika
python-jose
python-multipart
passlib[bcrypt]
hvac
